Реализует сохранение и обновление данных через localstorage

+++ src/api/provider.js
@@ -1,29 +1,79 @@
-export default class Provide {
-  constructor(api) {
+import Event from "../models/event.js";
+
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
+export default class Provider {
+  constructor(api, store) {
     this._api = api;
+    this._store = store;
   }
 
   getDestinations() {
-    return this._api.getDestinations();
+    if (isOnline()) {
+      return this._api.getDestinations();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getOptions() {
-    return this._api.getOptions();
+    if (isOnline()) {
+      return this._api.getOptions();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getEvents() {
-    return this._api.getEvents();
+    if (isOnline()) {
+      return this._api.getEvents()
+        .then((events) => {
+          events.forEach((event) => this._store.setItem(event.id, event.toRAW()));
+
+          return events;
+        });
+    }
+
+    const storeEvents = Object.values(this._store.getItems());
+
+    return Promise.reject(Event.parseEvents(storeEvents));
   }
 
   createEvent(event) {
-    return this._api.createEvent(event);
+    if (isOnline()) {
+      return this._api.createEvent(event);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   updateEvent(id, event) {
-    return this._api.updateEvent(id, event);
+    if (isOnline()) {
+      return this._api.updateEvent(id, event)
+        .then((newEvent) => {
+          this._store.setItem(newEvent.id, newEvent.toRAW());
+
+          return newEvent;
+        });
+    }
+
+    const localEvent = Event.clone(Object.assign(event, {id}));
+
+    this._store.setItem(id, localEvent.toRAW());
+
+    return Promise.reject(localEvent);
   }
 
   deleteEvent(id) {
-    return this._api.deleteEvent(id);
+    if (isOnline()) {
+      return this._api.deleteEvent(id)
+        .then(() => this._store.removeItem(id));
+    }
+
+    this._store.removeItem(id);
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 }
